Remove .toInt() for all param validators
no point in converting as express Request.params is always string so the conversion needs to be done in controller

+++ root/backend/app/validators/course-validator.ts
@@ -4,8 +4,7 @@ import validate from "../middlewares/validate";
 export const courseParamValidator: any = [
   param('courseId')
     .exists().withMessage("Missing courseId")
-    .isInt().withMessage("courseId must be a number")
-    .toInt(),
+    .isInt().withMessage("courseId must be a number"),
   validate,
 ]
