+++ root/backend/app/validators/intake-validator.ts
@@ -4,8 +4,7 @@ import validate from "../middlewares/validate";
 export const intakeParamValidator: any = [
   param('intakeId')
     .exists().withMessage("Missing intakeId")
-    .isInt().withMessage("intakeId must be a number")
-    .toInt(),
+    .isInt().withMessage("intakeId must be a number"),
   validate,
 ];
